revise vanilla and dictionary-based rotate90
